LPS-117566 Move to exception package
