+++ modules/apps/content-dashboard/content-dashboard-api/bnd.bnd
@@ -1,4 +1,6 @@
 Bundle-Name: Liferay Content Dashboard API
 Bundle-SymbolicName: com.liferay.content.dashboard.api
 Bundle-Version: 1.0.0
-Export-Package: com.liferay.content.dashboard.item.action
+Export-Package:\
+	com.liferay.content.dashboard.item.action,\
+	com.liferay.content.dashboard.item.action.exception
