+++ modules/apps/content-dashboard/content-dashboard-api/src/main/resources/com/liferay/content/dashboard/item/action/exception/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
